requirements.txt
python_decouple==3.0
dj_database_url==0.4.1
Django==1.9.5
django_rest_swagger==2.0.3
djangorestframework==3.4.4
psycopg2==2.6.2
gunicorn==19.6.0
whitenoise==3.2.1
django-rest-auth==0.8.1
djangorestframework-jwt==1.8.0
django-allauth==0.26.1
django-dropbox==0.0.2

